Fix date validation
